"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2-assignments/m6/ADS-2Week-1/Solution.java
@@ -19,7 +19,7 @@ class PageRank {
 			for (int j = 0; j < digr.V(); j++) {
 				for (int each : digr.adj(j)) {
 					if (each == i) {
-						temp += pray[j] / digr.outdegree(j);
+						temp += pray[j] / (double)digr.outdegree(j);
 					}
 				}
 			}
